Major edit, remove channel, fix encoding, simplify
Worker no longer has a GroupID nor takes a channel. It must also have
totalBits + 1 that are divisible by 6, for easy base64 encoding.
Formatting now removes bytes that won't be encoded based on the
totalBits, and removes the possible trailing A (empty 6 bits) when the
last byte will only have 2 bits encoded.

Worker now has a single ID, to allow that section to be more adjustable.
It also does not have a channel, because the channel was significantly
slower than just grabbing the NextID. Users of the library can use their
own channels if desired.

This simplification is to, in the future, allow shorter IDs to be
generated with a smaller timestamp value.